SDR_GUI/requirements.txt
pyserial
mttkinter
python-tk